requirements.txt
# LangGraph Wikimedia Curation Agent Requirements

# Core LangGraph and LangChain
langgraph>=0.0.55
langchain>=0.1.0
langchain-openai>=0.1.0

# Database
pymongo>=4.0.0

# HTTP and utilities
requests>=2.31.0
python-dotenv>=1.0.0

# Logging
coloredlogs>=15.0